Move storage package to seedu.navi package Change logo formatting
