+++ src/main/java/seedu/navi/canteenfinder/CanteenFinder.java
@@ -1,5 +1,11 @@
 package seedu.navi.canteenfinder;
 
+import seedu.navi.canteenfinder.landmark.Landmark;
+
+import java.util.ArrayList;
+
 // main class for canteen finder feature
 public class CanteenFinder {
+    public static final ArrayList<Landmark> LANDMARKS = new ArrayList<>();
+    
 }
