+++ src/main/java/seedu/navi/canteenfinder/landmark/Faculty.java
@@ -1,4 +1,4 @@
-package seedu.navi.canteenfinder.nus.landmark;
+package seedu.navi.canteenfinder.landmark;
 
 public class Faculty extends Landmark {
     public Faculty(String name) {
