+++ src/main/java/seedu/navi/canteenfinder/landmark/canteen/stall/StallCharacteristic.java
@@ -1,4 +1,4 @@
-package seedu.navi.canteenfinder.nus.landmark.canteen.stall;
+package seedu.navi.canteenfinder.landmark.canteen.stall;
 
 public class StallCharacteristic {
     private final boolean halalCertified;
