--- src/main/java/seedu/navi/canteenfinder/nus/NUS.java
@@ -1,15 +1,0 @@
-package seedu.navi.canteenfinder.nus;
-
-import seedu.navi.canteenfinder.nus.landmark.Landmark;
-import seedu.navi.canteenfinder.storage.Storage;
-
-import java.util.ArrayList;
-
-// this class will do the searching
-public class NUS {
-    public static final ArrayList<Landmark> LANDMARKS = new ArrayList<>();
-
-    public NUS() {
-        Storage.processDataFromFiles();
-    }
-}
